pyproject.toml
[project]
name = "norte-tech-portal"
version = "0.1.0"
description = "Portal Corporativo Norte Tech"
authors = [
    {name = "Luiz Fernando da Silva Guedes",email = "pitohuikun@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
